examples/azure-container-apps/pyproject.toml
[project]
name = "py-code-mode-azure-example"
version = "0.1.0"
description = "Azure Container Apps deployment example for py-code-mode"
requires-python = ">=3.12"
dependencies = [
    "py-code-mode[mcp]",
    "autogen-agentchat>=0.4",
    "autogen-ext[anthropic,azure]>=0.4",
    "azure-identity>=1.15.0",
    "python-dotenv",
    "tiktoken>=0.12.0",
    "redis>=5.0.0",
    "fastapi>=0.109.0",
    "pydantic>=2.0.0",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
py-code-mode = { path = "../..", editable = true }
